+++ scripts/ui/WidthButton.cs
@@ -17,31 +17,101 @@ partial class WidthButton : NButton
 
 	private VSlider _widthSlider;
 
+	private Label _widthLabel;
+
 	private HoverTip _hoverTip;
 
 	private Tween? _tween;
 
 	private bool _isOpen;
 
+	private bool _beOpenByCtrl;
+
 	private static readonly Color _activeColor = new Color(1f, 1f, 1f);
 
-	private static readonly Color _inactiveColor = new Color(0.3f, 0.3f, 0.3f);
+	private static readonly Color _inactiveColor = new Color(0.4f, 0.4f, 0.4f);
+
+	public WidthButton() : base()
+	{
+		Name = "WidthButton";
+		CustomMinimumSize = new Vector2(60, 60);
+		FocusNeighborLeft = "../ClearButton";
+		FocusNeighborRight = "../ColorButton";
+		FocusMode = FocusModeEnum.All;
+
+		_icon = new TextureRect(){
+            Name = "Icon",
+            Texture = GD.Load<Texture2D>("res://assets/BetterDrawing/PenNib.png"),
+            ExpandMode = TextureRect.ExpandModeEnum.IgnoreSize,
+            Size = new Vector2(50f, 50f),
+            Position = new Vector2(3f, 3f),
+            SelfModulate = new Color(0.3f, 0.3f, 0.3f),
+			MouseFilter = MouseFilterEnum.Ignore
+        };
+		_background = new ColorRect(){
+            Name = "Background",
+            Visible = false,
+            Color = new Color(0f, 0f, 0f, 0.8f),
+
+            AnchorBottom = 1f,
+            AnchorTop = 1f,
+            OffsetBottom = -60f,
+            OffsetTop = -469f,
+            OffsetLeft = 22f,
+            OffsetRight = 38f,
+			MouseFilter = MouseFilterEnum.Ignore
+        };
+		_widthSlider = new VSlider(){
+            Name = "WidthSlider",
+            Visible = false,
+
+            MinValue = 1f,
+            MaxValue = 32f,
+            Step = 0.5f,
+            Value = 4f,
+
+            AnchorBottom = 1f,
+            AnchorTop = 1f,
+            OffsetBottom = -60f,
+            OffsetTop = -469f,
+            OffsetLeft = 22f,
+            OffsetRight = 38f,
+        };
+		_widthSlider.Connect("value_changed", new Callable(this, nameof(OnWidthChanged)));
+		_widthLabel = new(){
+            Name = "WidthLabel",
+            Visible = false,
+            Text = "4.0",
+            GrowHorizontal = GrowDirection.Both,
+
+            AnchorBottom = 1f,
+            AnchorTop = 1f,
+            OffsetBottom = -469f,
+            OffsetTop = -492f,
+            OffsetLeft = 18f,
+            OffsetRight = 41f,
+			MouseFilter = MouseFilterEnum.Ignore
+        };
+        _widthLabel.AddThemeStyleboxOverride("normal", new StyleBoxFlat(){BgColor = new Color(0f, 0f, 0f, 0.8f)});
+		AddChild(_icon);
+        AddChild(_background);
+        AddChild(_widthSlider);
+        AddChild(_widthLabel);
+	}
 
 	public override void _Ready()
 	{
 		ConnectSignals();
 		_drawingToolHolder = (Control)GetParent();
-		_icon = GetNode<TextureRect>("Icon");
-		_background = GetNode<ColorRect>("Background");
-		_widthSlider = GetNode<VSlider>("WidthSlider");
-		_widthSlider.Connect("value_changed", new Callable(this, nameof(OnWidthChanged)));
-
 		NControllerManager.Instance.Connect(NControllerManager.SignalName.ControllerDetected, Callable.From(OnControllerUpdated));
 		NControllerManager.Instance.Connect(NControllerManager.SignalName.MouseDetected, Callable.From(OnControllerUpdated));
 		OnControllerUpdated();
 	}
 
-	private void OnWidthChanged(float newWidth){
+	private void OnWidthChanged(float newWidth)
+	{
+		DrawingDataAccess._widthMarker.UpdateWidth(newWidth + 18f);
+		_widthLabel.Text = newWidth.ToString("F1");
 		DrawingDataAccess._playerWidths[DrawingDataAccess._netService.NetId] = newWidth;
 		DrawingDataAccess._netService.SendMessage(new BetterDrawingMessage()
 		{
@@ -50,22 +120,38 @@ partial class WidthButton : NButton
 		});
 	}
 
-	protected override void OnPress()
+	protected override void OnRelease()
 	{
-		base.OnPress();
+		base.OnRelease();
 		OpenOrClose(!_isOpen);
 	}
 	
-	private void OpenOrClose(bool isOpen)
+	private void OpenOrClose(bool isOpen, bool changeFocus = true)
 	{
 		_isOpen = isOpen;
 		_widthSlider.Visible = isOpen;
 		_background.Visible = isOpen;
-		if (isOpen) {
-			OnUnfocus();
+		_widthLabel.Visible = isOpen;
+		if (changeFocus)
+		{
+			if (isOpen) OnUnfocus();
+			else OnFocus();
 		}
-		else {
-			OnFocus();
+	}
+
+	public void OpenOrCloseByCtrl(bool isOpen)
+	{
+		if (isOpen)
+		{
+			if (_isOpen) return;
+			_beOpenByCtrl = true;
+			OpenOrClose(true, false);
+		}
+		else
+		{
+			if (!_beOpenByCtrl) return;
+			_beOpenByCtrl = false;
+			OpenOrClose(false, false);
 		}
 	}
 
@@ -96,7 +182,7 @@ partial class WidthButton : NButton
 	private void OnControllerUpdated()
 	{
 		LocString description = new("better_drawing", "WIDTH_BUTTON.description");
-		LocString title = new("better_drawing", "WIDTH_BUTTON.title_mkb");
+		LocString title = (!NControllerManager.Instance.IsUsingController) ? new LocString("better_drawing", "WIDTH_BUTTON.title_mkb") : new LocString("better_drawing", "WIDTH_BUTTON.title_controller");
 		_hoverTip = new HoverTip(title, description);
 	}
 }